+++ src/main/java/com/example/Scrabble/Model/LocalServer/CommandFactory.java
@@ -1,6 +1,7 @@
-package com.example.Scrabble.Model;
+package com.example.Scrabble.Model.LocalServer;
 
 import com.example.Scrabble.Model.Game.GameManager;
+import com.example.Scrabble.Model.Player.GuestPlayer;
 
 public class CommandFactory {
     public Command createCommand(String request) {
