+++ src/main/java/com/example/Scrabble/Model/Player/HostPlayer.java
@@ -1,7 +1,8 @@
-package com.example.Scrabble.Model;
+package com.example.Scrabble.Model.Player;
 
 import com.example.Scrabble.Model.Game.GameManager;
-import com.example.Scrabble.ScrabbleServer.MyServer;
+import com.example.Scrabble.Model.LocalServer.PlayerHandler;
+import com.example.Scrabble.ServerUtils.MyServer;
 
 // import java.util.Random;
