+++ src/main/java/com/example/Scrabble/Model/Player/Player.java
@@ -1,4 +1,4 @@
-package com.example.Scrabble.Model;
+package com.example.Scrabble.Model.Player;
 
 public interface Player {
     public String getName();
