+++ src/main/java/com/example/Scrabble/ScrabbleDictionary/CacheManager/CacheReplacementPolicy.java
@@ -1,0 +1,6 @@
+package com.example.Scrabble.ScrabbleDictionary.CacheManager;
+
+public interface CacheReplacementPolicy{
+	void add(String word);
+	String remove(); 
+}
