--- src/main/java/com/example/Scrabble/Server/CacheManager/CacheReplacementPolicy.java
@@ -1,6 +1,0 @@
-package com.example.Scrabble.Server.CacheManager;
-
-public interface CacheReplacementPolicy{
-	void add(String word);
-	String remove(); 
-}
