+++ src/main/java/com/example/Scrabble/ServerUtils/MyServer.java
@@ -1,4 +1,4 @@
-package com.example.Scrabble.ScrabbleServer;
+package com.example.Scrabble.ServerUtils;
 
 
 import java.io.IOException;
