+++ src/main/java/com/example/Scrabble/Test/GuestPlayerTest.java
@@ -1,6 +1,6 @@
 package com.example.Scrabble.Test;
 
-import com.example.Scrabble.Model.GuestPlayer;
+import com.example.Scrabble.Model.Player.GuestPlayer;
 
 import static java.lang.Thread.sleep;
