+++ src/test/java/com/example/Scrabble/Model/PlayerHandlerTest.java
@@ -1,8 +1,9 @@
 package com.example.Scrabble.Model;
 
+import com.example.Scrabble.Model.LocalServer.PlayerHandler;
 import org.junit.jupiter.api.Test;
 
-import com.example.Scrabble.ScrabbleServer.MyServer;
+import com.example.Scrabble.ServerUtils.MyServer;
 
 import static org.junit.jupiter.api.Assertions.*;
